Split up controller into multiple files
