+++ controller/loginController.js
@@ -1,0 +1,36 @@
+var express = require("express");
+var db = require("../models");
+var passport = require("../config/passport");
+
+var router = express.Router();
+
+//LOGIN ROUTES======================================================================================================
+
+//logging in route
+router.post("/api/login/", passport.authenticate("local"), function(req, res) {
+    res.json("/members");
+});
+
+//signing up account route
+router.post("/api/signup", function(req, res) {
+
+    db.User.create({
+        username: req.body.username,
+        password: req.body.password
+    }).then(function() {
+        res.redirect(307, "/api/login");
+    }).catch(function(err) {
+        console.log("Getting error");
+        res.json(err);
+        
+    });
+});
+
+
+// Route for logging user out
+router.get("/logout", function(req, res) {
+    req.logout();
+    res.redirect("/");
+});
+
+module.exports = router;
