+++ models/favorite.js
@@ -4,17 +4,15 @@ module.exports = function(sequelize, DataTypes) {
 
     });
 
-
-
+    Favorite.associate = function(models) {
+        // Associating Users with Meals through Junction Tables Faves and Days
+        //models.Favorite.hasMany(models.Meal);
+        models.Favorite.belongsTo(models.Meal);
+        models.Favorite.belongsTo(models.User);
+    
+    
+    };
 
     return Favorite;
 };
 
-// Favorite.associate = function(models) {
-//     // Associating Users with Meals through Junction Tables Faves and Days
-//     //models.Favorite.hasMany(models.Meal);
-//     models.Favorite.belongsTo(models.Meal);
-//     models.Favorite.belongsTo(models.User);
-
-
-// };
